requirement.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
pytz==2023.3.post1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
typing_extensions==4.7.1
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.3.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.5
django-compat==1.0.15
django-cors-headers==4.3.0
django-import-export==3.3.1
django-sslserver==0.22
django4-background-tasks==1.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docopt==0.6.2
ecdsa==0.18.0
enum-compat==0.0.3
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.0
future==0.18.3
idna==3.4
intuit-oauth==1.2.4
ipython==8.16.1
jedi==0.19.1
MarkupPy==1.14
matplotlib-inline==0.1.6
numpy==1.26.1
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pipreqs==0.4.13
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-jose==3.3.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
rsa==4.9
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tablib==3.5.0
traitlets==5.11.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.6
wcwidth==0.2.8
xlrd==2.0.1
xlwt==1.3.0
xmldict==0.4.1
xmltodict==0.13.0
yarg==0.1.9